# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: CodeForge API
  description: |
    REST API for CodeForge -- AI coding agent orchestration platform.

    All endpoints under /api/v1 require a valid JWT bearer token unless
    explicitly marked as public (login, setup-status, health).

    Refresh tokens are delivered as httpOnly cookies on the /auth path.
  version: "0.8.0"
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: http://localhost:8080
    description: Local development (root, for /health)
...
paths:
...
  /runs/{id}/trajectory:
    get:
      tags: [Runs]
      summary: Get the trajectory (step-by-step log) of a run
      description: Returns the full sequence of agent steps, tool calls, and results.
      operationId: getTrajectory
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RunID"
      responses:
        "200":
          description: Trajectory data
          content:
            application/json:
              schema:
                type: object
                properties:
                  run_id:
                    type: string
                  steps:
                    type: array
                    items:
                      type: object
                      properties:
                        step:
                          type: integer
                        type:
                          type: string
                        tool_name:
                          type: string
                        input:
                          type: string
                        output:
                          type: string
                        timestamp:
                          type: string
                          format: date-time
        "404":
          $ref: "#/components/responses/NotFound"

  # -----------------------------------------------
  # Settings
  # -----------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from POST /auth/login.
        Pass as: Authorization: Bearer <token>
  parameters:
...
    RunID:
      name: id
      in: path
      required: true
      description: Run ID
      schema:
        type: string
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...